The initial KAPRIEN API
Implement following issue #1

This is the initial version and is compatible with the current
kaprien-cli.
kaprien-cli can generates the initial payload that contains the Kaprien
server settings and the initial delegations (Targets -> bin and bin ->
bins)

The Kaprien Server (API) is able to bootstrap the server, lock the
Metadata to avoid another initialization, store the Metadata to the
storage (local) and also the credentials to the key storage (local).

The initial version contains the API documentation and basic
documentation for development.

The unit tests is missing and will be followed for another issue. --
new implementations will not be done without tests in the same commit as
done here.

Signed-off-by: Kairo de Araujo <kairo@kairo.eti.br>