Implemented endpoint to retrieve the metadata
This commit implements the endpoint `metadata` with the method `GET`
thar retrives the Repository Metadata from the Storage.
It retrieves the latest Metadata for each role.

An optional parameter is available 'rolename' if the system needs to
retrive only one specific role metadata.

The TUFMetadata data structure used by API was moved from the
`bootstrap` to `utils` one it is a common Object used in the API.

Some fixes in the tox, added new functionalities in the Makefile.

Closes: #5
Closes: #6

Signed-off-by: Kairo de Araujo <kairo@kairo.eti.br>