Include the kaprien-mq container in development
This commit introduces the kaprien-mq as a RabbitMQ container that will
be used directly by kaprien-rest-api.
Any action to the Kaprien TUF Metadata will not be handled directly by
the kaprien-rest-api in the metadata, but it will be sent to a message
queue.
The actions in the TUF Metadata will be handled by another service named
kaprien-worker.
It avoids race conditions and enables the scalability of the deployment.

This service is now handled only in development environment.

Signed-off-by: Kairo de Araujo <kairo@kairo.eti.br>