Implement a new endpoint to add targets
This commit implements the endpoint `/api/v1/targets` with method `POST`
to add new targets to the Metadata Repository.
This implementation adds the endpoint, documentation for the required
JSON payload and the response body.

This commit does not submit the targets to the message queue, it needs
to be implemented by #39.

A small fix in the development `docker-compose.yml` was added in order
to refresh automatically in case of changes in the `app.py`

Signed-off-by: Kairo de Araujo <kairo@kairo.eti.br>