Build Docker Image and publishes to GPR
CI that build the docker image and publishes to the Github Packages
Repositories (GPR), Docker Registry.
This is the development versions, latest version of `main` Branch.

In case of pull request it only runs the build to test the Docker Build
In case of push, it will publish a new dev in the GPR

Closes #45

Signed-off-by: Kairo de Araujo <kairo@kairo.eti.br>