Fix security issue with Tokens when empty scope
The initial design for the built-in Authentication/Authorization was
based on multiple users with scopes.
Users, after authentication, would receive a token with all its
defined scopes.

Authentication and Authorization are not core components of RSTUF.
We simplified this design to have only one user (admin), and the
admin can generate Tokens with specific Scopes.
During this design change, we did not change the user fetching all
scopes if the scope is empty/None.
We even want to allow the user to disable the
Authentication/Authorization layer (see
https://github.com/vmware/repository-service-tuf/issues/41) and
use their  Authentication/Authorization layer.

- Add scope as required
- Add tests to avoid regression
- Hide password in the swagger try out
- Gives a clearer view of the scopes in the documentation (swagger)

Closes #175

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>