docs/tests: add missing delegation path in payload
In the example payload.json we use for our swagger.json we define the
metadata structure.
There we define that "2.targets.json" delegates all target files
within the following paths "*/*" and "*/*/*/*".
I noticed we are missing the case where there is a target file in the
root directory or in "*".
That's a case we want to test in our functional tests in the umbrella
repository and it makes sense to support such targets.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>