Add an option to add task_id to target custom
Add an option which can be used for debugging to add the "task_id"
of the process that adds the target into the "custom" attributes of
all targets.
By default this option is enabled as it will add a small overhead, but
can be quite useful for debugging in a bad scenario.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>