Add publish_targets field in add/remove targets
Add an optional field called "publish_targets" in the add/remove
targets API calls. If no value is provided or it's set to true,
publishing of targets will happen the same as now: a subtask will
be started that on completion will have published the modified targets.
If it's set to false, publishing of the targets file changes won't be
published (meaning there won't be any changes in the metadata files
as well) until specifically requested with the newly introduced
API call "api/v1/targets/publish".

This feature is was originated in Task:
https://github.com/vmware/repository-service-tuf-worker/issues/175
discussion

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>