Update expected payload in post/delete tests
The new fields "publish_targets" and "add_task_id_to_custom" are defined
as optional, but with a default value meaning even if we don't provide
value for them in the payload.json they will be defined with the given
default value for them in repository_service_api/targets.py.
That's why the tests should be updated to provide value always for
those fields.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>