Create an auth service for the rstuf api
As it mentioned in
https://github.com/vmware/repository-service-tuf/issues/41
the API service has a simple built-in authentication and
authorization implementation for small deployments.

We want to be able to disable this built-in feature and add
support of other methods.

This commit is a first step on this direction. We move the
authentication code to a new custom authentication service
inside of the API service.

We achieve separation of concerns and we can easily replace
the custom authentication service with another one without
touching the business logic, if we have an authentication
port with the same methods implemented for every different
authentication type (e.g., custom service, aws api gateway).

This implementation helps us to avoid mocking, and replace
the CustomAuthenticationService with a TestAuthenticationService
when we are on the testing environment. The test service must
implement the same methods, but don't need to access any db.
It can store objects in memory.

Signed-off-by: Michalis Kapsalakis <kapsalis1989@gmail.com>