fix: env variable for custom Redis result DB
This commit fixes the environment variable used to setup a custom
Redis backend result database ID, as documented in the API Guide
documentation for the container configuration.

It was using by mistake the environment variable used to setup the
custom Redis repository settings ID for the container.

It impacts the RSTUF deployment for custom scenarios.

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>