More flexible disabling endpoint feature
This feature gives better flexibility to the RSTUF user to disable
endpoints.

The user can disable specific endpoints and methods in a simple
configuration using the `RSTUF_DISABLE_ENDPOINTS`.

It reduces the number of settings parameters (environment variables)
`RSTUF_BOOTSTRAP_NODE`, and `RSTUF_TOKEN_NODE`, making the feature
more flexible as we add more endpoints and methods to the RSTUF API.

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>