feat: bootstrap distributed async sign
It implements the base structure endpoints for the distributed async
signing for the RSTUF.

The implementation also adds support to the distributed async sign
for the bootstrap. The user can initiate the bootstrap ceremony
process with the Root metadata not fully signed, and the other root
ey owners can start the signing process.

It requires
 - https://github.com/repository-service-tuf/repository-service-tuf-worker/issues/300

Closes #355

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>