Add BDD tests for Token Feature to RSTUF API
The current BDD Tokens tests are part of integration tests in the
RSTUF umbrella repository.
This BDD feature test doesn't belong to the integration tests once
API Token authentication/authorization (auth) feature is a single
component.

The BDD test is still valid once it describes very well the expected
behavior.

This commit adds the test to the API and is part of the
https://github.com/repository-service-tuf/repository-service-tuf/issues/41

Closes #281

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>