Add a new DELETE /api/v1/metadata/sign endpoint
Add a new DELETE /api/v1/metadata/sign endpoint.
The purpose of this endpoint is to give a direct way for repository
administrators to stop a DAS ceremony.
This DAS process can be started from both POST /api/v1/bootstrap and
POST /api/v1/metadata.
This will be different than the suggested PATCH /api/v1/metadata/sign
endpoint as this will give the opportunity to:
- stop the DAS process immediately and do not risk other maintainers
signing the metadata and passing the threshold requirement.
- take it slow and decide what the new bootstrap or metadata update
should look like

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>