Fix bootstrap_state checks in two places
Fix the bootstrap_state checks in `post_bootstrap` and `get_bootstrap`.
This makes sure that we say there is a metadata when:
- bootstrap is executed in the moment ("pre" state)
- bootstrap is in a signing process ("signing" state)

Also, fix some small things like "task_id" not sent when calling
"get_bootstrap".

Added unit tests for "get_bootstrap" to make sure this mistake
can be easily noticed if it happens again, for "post_bootstrap" there
are tests already implemented.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>