Refactor DELETE /api/v1/metadata/sign
Add a refactoring of `DELETE /api/v1/metadata/sign` with body payload
to a `POST /api/v1/metadata/sign/delete`.

It makes the RSTUF API more compatible with OpenAPI and HTTP
specifications, avoiding the body in the DELETE method.

Also, the current implementation is asynchronous, which makes more
semantical adding a post as the resource is not deleted instantly

Closes #464

Signed-off-by: Kairo de Araujo <kdearaujo@vmware.com>