Feature: support to unrecognized fields (Root)
This commit adds the support to unrecognized fields to the Root.

The Root.signed.unrecognized_fields can be used to store relevant
information in the TUF metadata used by TUF clients or even RSTUF
Worker.

It is added to the `TUFSigned`, allowing extra fields.
The unrecognized fields must to have the format `x-<vendor>-<name>`

We add a pydantic `root_validator` which will verify all extra fields if
they follow the required format.

Part of #608

Signed-off-by: Kairo Araujo <kairo.araujo@testifysec.com>