Update payloads and keys for new admin2 CLI
Major refactoring is happening in the CLI where we have moved towards
a new set of key formats that are no longer specific to
securesystemslib.
As a part of this change new keys were introduced in the CLI and we
need to update the keys in the CLI as well.

Additionally, new ceremonies, metadata updates and sign payloads were
generated. Also, I made sure we don't have additional spaces in our
JSON files.

Signed-off-by: Martin Vrachev <martin.vrachev@broadcom.com>